+++ bazaar/kissfir/kissfir.cpp
@@ -76,7 +76,7 @@ CONSOLE_APP_MAIN
     }
     fseek(filtfile,0,SEEK_END);
     nh = ftell(filtfile) / sizeof(kffsamp_t);
-    if (verbose) fprintf(stderr,"%d samples in FIR filter\n",nh);
+    if (verbose) fprintf(stderr,"%d samples in FIR filter\n",(int)nh);
     h = (kffsamp_t*)malloc(sizeof(kffsamp_t)*nh);
     fseek(filtfile,0,SEEK_SET);
     fread(h,sizeof(kffsamp_t),nh,filtfile);
